requirements.dev.txt
# dev dependencies
ipython
ipdb
pudb
pytest
pytest-watch


<<<<<<< Updated upstream
# auto reload
django-livereload
django-livereload-server
=======
# auto reload front-end
django-browser-reload
>>>>>>> Stashed changes


# Code Quality
flake8
pyproject-flake8
black
isort
